backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
websockets>=14.0
anthropic>=0.43
deepgram-sdk>=3.9
elevenlabs>=1.19
aiosqlite>=0.20
python-dotenv>=1.0
textblob>=0.18
